+++ internal/examples/pattern.go
@@ -1,4 +1,4 @@
-package doc
+package examples
 
 import (
 	"regexp/syntax"